Switch everything over to the pixi map preview, delete the original map preview Fix some lingering start box bugs Slightly optimize redraw
