+++ src/renderer/components/battle/BattlePreview.vue
@@ -1,12 +1,7 @@
 <template>
     <div class="flex-col gap-md fullheight">
-        <MapPreview
-            :map="mapScriptName"
-            :isSpectator="true"
-            :myTeamId="0"
-            :startBoxes="startBoxes"
-            :startPosType="startPosType"
-            :startPositions="startPositions"
+        <MapOverviewCard
+            :map="map"
         />
 
         <div class="teams">
@@ -14,8 +9,9 @@
                 <div class="team-title">Players</div>
                 <div class="contenders">
                     <template v-if="isBattle(battle)">
-                        <template v-for="(contender, i) in battle.contenders.value" :key="`contender${i}`">
-                            <BattlePreviewParticipant :contender="contender" />
+                        <template v-for="(contender, i) in battle.contenders.value"
+                                  :key="`contender${i}`">
+                            <BattlePreviewParticipant :contender="contender"/>
                         </template>
                     </template>
                     <template v-else-if="isReplay(battle)">
@@ -75,7 +71,7 @@ import trophyVariant from "@iconify-icons/mdi/trophy-variant";
 import { groupBy } from "jaz-ts-utils";
 import { computed, ComputedRef } from "vue";
 
-import MapPreview from "@/components/maps/MapPreview.vue";
+import MapOverviewCard from "@/components/maps/MapOverviewCard.vue";
 import BattlePreviewParticipant from "@/components/misc/BattlePreviewParticipant.vue";
 import { AbstractBattle } from "@/model/battle/abstract-battle";
 import { StartBox, StartPosType } from "@/model/battle/battle-types";
@@ -87,8 +83,10 @@ const props = defineProps<{
     showSpoilers?: boolean;
 }>();
 
-const mapScriptName = computed(() => {
-    return props.battle instanceof AbstractBattle ? props.battle.battleOptions.map : props.battle.mapScriptName;
+const map = computed(() => {
+    return props.battle instanceof AbstractBattle
+        ? api.content.maps.getMapByScriptName(props.battle.battleOptions.map)
+        : api.content.maps.getMapByScriptName(props.battle.mapScriptName);
 });
 const gameVersion = computed(() =>
     props.battle instanceof AbstractBattle ? props.battle.battleOptions.gameVersion : props.battle.gameVersion
