+++ src/renderer/components/controls/Range.vue
@@ -21,7 +21,6 @@ export interface Props extends SliderProps {
 const props = defineProps<Props>();
 const emits = defineEmits<{
     (event: "update:modelValue", value: number | number[]): void
-    (event: "onUpdate", value: number | number[]): void
 }>();
 
 const slider: Ref<null | Props> = ref(null);
@@ -36,13 +35,11 @@ onMounted(() => {
 
 function onSlide(input: number | number[]) {
     emits("update:modelValue", input);
-    emits("onUpdate", input);
 }
 
 function onInput(input: number | number[]) {
     if (typeof input === "number" && !Array.isArray(input)) {
         emits("update:modelValue", input);
-        emits("onUpdate", input);
     }
 }
 </script>
