+++ src/renderer/components/maps/MapPreview.vue
@@ -1,219 +1,229 @@
 <template>
-    <div ref="mapPreviewEl" class="map-preview">
-        <canvas ref="canvas" class="canvas" />
-    </div>
+    <div ref="canvasContainerEl" class="canvas-container"></div>
 </template>
 
 <script lang="ts" setup>
-import { entries, SignalBinding } from "jaz-ts-utils";
-import { computed, onMounted, onUnmounted, Ref, ref, watch, WatchStopHandle } from "vue";
+import { rgb2hex } from "@pixi/utils";
+import { useElementSize } from "@vueuse/core";
+import { Application, Graphics, Sprite } from "pixi.js";
+import { onMounted, onUnmounted, ref, watch } from "vue";
 
 import { StartBox, StartPosType } from "@/model/battle/battle-types";
-
-type Transform = { x: number; y: number; width: number; height: number };
+import { MapData } from "@/model/map-data";
+import { CurrentUser } from "@/model/user";
+import { mipmapSize } from "@/workers/parse-map";
 
 const props = defineProps<{
-    map: string;
+    map?: MapData;
+    currentUser?: CurrentUser;
     startPosType?: StartPosType;
     startBoxes?: Record<number, StartBox | undefined>;
-    myTeamId?: number;
-    isSpectator?: boolean;
     startPositions?: Array<
         | {
-              position: { x: number; z: number };
-              rgbColor?: { r: number; g: number; b: number };
-          }
+        position: { x: number; z: number };
+        rgbColor?: { r: number; g: number; b: number };
+    }
         | undefined
     >;
 }>();
 
-const mapPreviewEl: Ref<HTMLDivElement | null> = ref(null);
-const canvas: Ref<HTMLCanvasElement | null> = ref(null);
-let context: CanvasRenderingContext2D;
+const app = ref<Application>();
+const canvasContainerEl = ref<HTMLElement>();
+const parentSize = useElementSize(canvasContainerEl);
 
-const mapData = computed(() => api.content.maps.installedMaps.find((map) => map.scriptName === props.map));
-const mapImages = computed(() => api.content.maps.getMapImages(mapData.value));
-
-let watchStopHandle: WatchStopHandle | undefined;
-let mapCachedSignalBinding: SignalBinding | undefined;
-
-onMounted(async () => {
-    if (!canvas.value || !mapPreviewEl.value) {
-        return;
-    }
-
-    const canvasWidth = await onCanvasResize();
-
-    canvas.value.width = canvasWidth;
-    canvas.value.height = canvasWidth;
-
-    context = canvas.value.getContext("2d")!;
-
-    await loadMap();
-
-    watchStopHandle = watch(
-        [() => props.map, () => props.startPosType, () => props.startBoxes, () => props.myTeamId],
-        async () => {
-            await loadMap();
-        },
-        { deep: true }
-    );
-
-    mapCachedSignalBinding = api.content.maps.onMapCached.add(async (data) => {
-        if (data.scriptName === props.map) {
-            mapData.effect.run();
-            mapImages.effect.run();
-            await loadMap();
-        }
-    });
-});
-
+onMounted(setup);
 onUnmounted(() => {
-    watchStopHandle?.();
-    watchStopHandle = undefined;
-
-    mapCachedSignalBinding?.destroy();
-    mapCachedSignalBinding = undefined;
+    // make sure pixi app is properly destroyed after out transition has finished
+    setTimeout(() => {
+        app.value?.destroy(true);
+    }, 150);
 });
 
-function onCanvasResize() {
-    return new Promise<number>((resolve) => {
-        const resizeObserver = new ResizeObserver((resizeObserverEntries) => {
-            const rect = resizeObserverEntries[0].contentRect;
+watch([parentSize.width, parentSize.height], ([width, height], [oldWidth, oldHeight]) => {
+    if (app.value && width && height) {
+        const smallestDimension = Math.min(width, height);
+        app.value.renderer.resize(smallestDimension, smallestDimension);
+        app.value.render();
+        onResize();
+    }
+});
 
-            resolve(Math.max(rect.width, rect.height));
-            resizeObserver.disconnect();
-        });
-        resizeObserver.observe(canvas.value!);
+watch(() => props.map, setMapImage);
+watch(() => props.startBoxes, drawBoxes);
+watch(() => props.startPositions, drawStartPositions);
+watch(
+    () => props.currentUser?.battleStatus.teamId,
+    () => {
+        drawBoxes();
+        drawStartPositions();
+    }
+);
+watch(() => props.currentUser?.battleStatus.isSpectator, drawBoxes);
+watch(
+    () => props.startPosType,
+    () => {
+        drawBoxes();
+        drawStartPositions();
+    }
+);
+
+let mapSprite: Sprite | undefined;
+let boxesGfx: Graphics | undefined;
+let startPositionsGfx: Graphics | undefined;
+
+function setup() {
+    app.value = new Application({
+        background: "#000",
+        backgroundAlpha: 0.3,
+        antialias: true,
     });
+    if (app.value.view && app.value.view instanceof HTMLCanvasElement) {
+        canvasContainerEl.value?.appendChild(app.value.view);
+    }
+
+    boxesGfx = new Graphics();
+    app.value.stage.addChild(boxesGfx);
+
+    startPositionsGfx = new Graphics();
+    app.value.stage.addChild(startPositionsGfx);
+
+    setMapImage();
 }
 
-async function loadMap() {
-    if (!canvas.value) {
+function onResize() {
+    // this is often 0 in modals, don't want to waste render time
+    if (parentSize.width.value > 0 && parentSize.height.value > 0) {
+        resizeMapImage();
+        drawBoxes();
+        drawStartPositions();
+    }
+}
+
+async function setMapImage() {
+    if (!app.value || !props.map) {
         return;
     }
 
-    const canvasWidth = canvas.value?.width;
+    if (mapSprite) {
+        app.value.stage.removeChild(mapSprite);
+    }
 
-    let mapTransform: Transform = { x: 0, y: 0, width: canvasWidth, height: canvasWidth };
+    const textureImage = api.content.maps.getMapImages(props.map).textureImagePath;
+    mapSprite = Sprite.from(textureImage, {
+        width: props.map.width * mipmapSize * 16,
+        height: props.map.height * mipmapSize * 16,
+    });
+    app.value.stage.addChildAt(mapSprite, 0);
 
-    const textureMap = await loadImage(mapImages.value.textureImagePath);
+    onResize();
+}
 
-    const widthToHeightRatio = textureMap.width / textureMap.height;
-    if (widthToHeightRatio > 1) {
-        mapTransform.height = mapTransform.height / widthToHeightRatio;
+function resizeMapImage() {
+    if (!app.value || !mapSprite) {
+        return;
+    }
+
+    if (mapSprite.width > mapSprite.height) {
+        mapSprite.width = app.value.view.width;
+        mapSprite.scale.y = mapSprite.scale.x;
     } else {
-        mapTransform.width = mapTransform.width * widthToHeightRatio;
+        mapSprite.height = app.value.view.height;
+        mapSprite.scale.x = mapSprite.scale.y;
     }
-
-    mapTransform.x = (canvasWidth - mapTransform.width) / 2;
-    mapTransform.y = (canvasWidth - mapTransform.height) / 2;
-
-    mapTransform = roundTransform(mapTransform);
-
-    context.clearRect(0, 0, canvasWidth, canvasWidth);
-
-    context.drawImage(textureMap, mapTransform.x, mapTransform.y, mapTransform.width, mapTransform.height);
-
-    drawStartPosType(mapTransform);
+    mapSprite.x = app.value.view.width * 0.5 - mapSprite.width * 0.5;
+    mapSprite.y = app.value.view.height * 0.5 - mapSprite.height * 0.5;
 }
 
-function drawStartPosType(mapTransform: Transform) {
-    switch (props.startPosType) {
-        case StartPosType.Boxes:
-            drawBoxes(mapTransform);
-            break;
-        case StartPosType.Fixed:
-        case StartPosType.Random:
-            drawFixedPositions(mapTransform);
-            break;
+function drawBoxes() {
+    if (!app.value || !boxesGfx || !mapSprite) {
+        return;
     }
-}
 
-function drawFixedPositions(mapTransform: Transform) {
-    if (mapData.value?.startPositions) {
-        for (const startPos of mapData.value.startPositions) {
-            drawFixedPosition(mapTransform, startPos);
+    if (props.startPosType !== StartPosType.Boxes) {
+        boxesGfx.visible = false;
+        return;
+    } else {
+        boxesGfx.visible = true;
+        boxesGfx.alpha = 0.2;
+        boxesGfx.clear();
+    }
+
+    const isSpectator = props.currentUser?.battleStatus.isSpectator;
+    const myTeamId = props.currentUser?.battleStatus.teamId;
+
+    for (const teamIdStr in props.startBoxes) {
+        const box = props.startBoxes[teamIdStr];
+        if (!box) {
+            continue;
         }
+
+        const teamId = parseInt(teamIdStr);
+        if (isSpectator) {
+            boxesGfx?.beginFill(0xffffff);
+        } else if (myTeamId === teamId) {
+            boxesGfx?.beginFill(0x00ff00);
+        } else {
+            boxesGfx?.beginFill(0xff0000);
+        }
+
+        boxesGfx?.drawRect(
+            box.xPercent * mapSprite.width + mapSprite.x,
+            box.yPercent * mapSprite.height + mapSprite.y,
+            box.widthPercent * mapSprite.width,
+            box.heightPercent * mapSprite.height
+        );
     }
 }
 
-function drawFixedPosition(mapTransform: Transform, position: { x: number; z: number }, color = "rgba(255, 255, 255, 0.6)") {
-    if (!mapData.value) return;
+function drawStartPositions() {
+    if (!app.value || !mapSprite || !startPositionsGfx || !props.map) {
+        return;
+    }
 
-    const xPos = mapTransform.x + mapTransform.width * (position.x / (mapData.value.width * 512));
-    const yPos = mapTransform.y + mapTransform.height * (position.z / (mapData.value.height * 512));
+    if (props.startPosType === StartPosType.Boxes) {
+        startPositionsGfx.visible = false;
+        return;
+    } else {
+        startPositionsGfx.visible = true;
+        startPositionsGfx.clear();
+    }
 
-    context.fillStyle = color;
-    context.beginPath();
-    context.arc(xPos, yPos, 5, 0, 2 * Math.PI);
-    context.fill();
-    context.closePath();
-}
-
-function drawBoxes(mapTransform: Transform) {
-    if (!props.startBoxes) return;
-
-    const startBoxEntries = entries(props.startBoxes);
-    startBoxEntries.forEach(([id, box], i) => {
-        if (box) {
-            if (props.isSpectator) {
-                context.fillStyle = "rgba(255, 255, 255, 0.2)";
-            } else if (props.myTeamId === i) {
-                context.fillStyle = "rgba(0, 255, 0, 0.2)";
-            } else {
-                context.fillStyle = "rgba(255, 0, 0, 0.2)";
-            }
-
-            context.strokeStyle = "rgba(255, 255, 255, 0.5)";
-            context.lineWidth = 1;
-
-            let boxTransform: Transform = {
-                x: mapTransform.x + mapTransform.width * box.xPercent,
-                y: mapTransform.y + mapTransform.height * box.yPercent,
-                width: mapTransform.width * box.widthPercent,
-                height: mapTransform.height * box.heightPercent,
+    const startPositions =
+        props.startPositions ??
+        props.map.startPositions?.map((pos) => {
+            return {
+                position: pos,
+                rgbColor: { r: 255, g: 255, b: 255 },
             };
-            boxTransform = roundTransform(boxTransform);
-            context.fillRect(boxTransform.x, boxTransform.y, boxTransform.width, boxTransform.height);
+        }) ??
+        [];
+
+    for (const startPos of startPositions) {
+        if (!startPos) {
+            continue;
         }
-    });
-}
 
-function loadImage(url: string) {
-    return new Promise<HTMLImageElement>((resolve) => {
-        const img = new Image();
-        img.onload = () => resolve(img);
-        img.src = url;
-    });
-}
+        let color = 0xffffff;
+        if (startPos.rgbColor) {
+            color = rgb2hex([startPos.rgbColor.r / 255, startPos.rgbColor.g / 255, startPos.rgbColor.b / 255]);
+        }
+        startPositionsGfx.beginFill(color);
 
-function roundTransform(transform: Transform) {
-    return {
-        x: Math.floor(transform.x),
-        y: Math.floor(transform.y),
-        width: Math.ceil(transform.width),
-        height: Math.ceil(transform.height),
-    };
+        const x = (startPos.position.x / (props.map.width * 512)) * mapSprite.width + mapSprite.x;
+        const y = (startPos.position.z / (props.map.height * 512)) * mapSprite.height + mapSprite.y;
+        const radius = 5;
+        startPositionsGfx.drawEllipse(x - radius * 0.5, y - radius * 0.5, radius, radius);
+    }
 }
 </script>
 
 <style lang="scss" scoped>
-.map-preview {
+.canvas-container {
+    overflow: hidden;
+    width: 100%;
     display: flex;
-    background: rgba(0, 0, 0, 0.3);
-    border: 1px solid rgba(255, 255, 255, 0.1);
-    position: relative;
     align-items: center;
     justify-content: center;
-    max-width: 100%;
-    max-height: 100%;
-}
-.canvas {
-    margin: 0;
-    aspect-ratio: 1;
-    width: 100%;
-    max-width: 100%;
-    max-height: 100%;
+    border: 1px solid rgba(255, 255, 255, 0.1);
 }
 </style>
